requirements.txt
pandas>=2.0
python-docx>=0.8.11
pyyaml>=6.0
python-dotenv>=1.0
tiktoken>=0.5
anthropic>=0.18
chromadb>=0.4
playwright>=1.40
pypdf>=4.0
sentence-transformers>=2.2
numpy>=1.24
flask>=3.0
